Adding Code + Notes and Stuff
This is a preliminary code, containing most of the logic necessary for
making a shell. However, this code is untested and may contain bugs.
Features that need to be implemented should all be marked with a TODO
and additional encapsulation may be possible. Required features include
trimming cstrings and parsing specific words from commands.